cleanup: move distance_string() to string-format.cpp
A long term project to split up the qthelper.h header.
